+++ zcash_transparent/src/lib.rs
@@ -6,9 +6,13 @@
 
 #![no_std]
 
+#[cfg(feature = "std")]
+extern crate std;
+
 pub mod address;
 pub mod builder;
 pub mod bundle;
+pub mod coinbase;
 pub mod keys;
 pub mod pczt;
 pub mod sighash;