+++ components/zcash_protocol/CHANGELOG.md
@@ -10,6 +10,11 @@ workspace.
 
 ## [Unreleased]
 
+## [0.7.2] - 2025-12-10
+
+### Added
+- `zcash_protocol::txid::TxId::NULL`
+
 ## [0.7.1] - 2025-10-18
 
 ### Fixed